fix(headless): update the rga events to send the searchId instead of the streamId as response id (#4766)
## [SFINT-5854](https://coveord.atlassian.net/browse/SFINT-5854)

We want to send the searchId with all the RGA EP events so every event
can properly be tied back to a search event when processing the
translations EP → UA events.

We were previously sending the `stream id` as a response id in the RGA
events, but we figured that will this cause for us issue when it comes
to trying to tie RGA events to the search events that were made with it
so we are doing this update.



https://github.com/user-attachments/assets/a6e3e029-1e28-4576-8b36-7b04414a3c5f



[SFINT-5854]:
https://coveord.atlassian.net/browse/SFINT-5854?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ