feat(headless commerce): add commerce parameters slice and excludeDefaultParameters controller option (#4759)
https://coveord.atlassian.net/browse/KIT-3462

We have been struggling with the implementation of the parameter manager
in the headless commerce SSR use case, mostly because the way the
parameter manager sub-controller currently works makes it so that its
state will typically change after we fetch the initial state (unless
we're restoring parameters from the URL) as the "default" parameters
(e.g., the page, perPage, and sortCriteria values configured for the
page being fetched) will be returned by the CAPI.

This PR adds a new commerce parameters state section / reducer, which
allows the controller to only keep track of parameters modified through
dispatched actions (i.e., through user interactions).

This would be a breaking change, so we're introducing this as an option
on the commerce parameters sub-controller which will default to false in
the non-SSR sub-package, and will default to true in the yet to be
released SSR sub-package.

A few things worth mentioning that this PR also modifies:

- The commerce pagination slice now resets the `page` to its initial
value (0) when the `restoreSearchParameters` or
`restoreProductListingParameters` action is dispatched and the `page`
value is undefined in the action payload. It also resets `page` to `0`
when `setPageSize` is dispatched (see
https://github.com/coveo/ui-kit/pull/4759/files#diff-3418e45cea101e732b2ca21969a80f152654268fa3334c2e3a369f95e2d172a5).

- Similarly, the commerce sort slice now resets the `sortCriteria` to
its initial value (relevance) when restoring parameters if
`sortCriteria` is undefined in the action payload (see
https://github.com/coveo/ui-kit/pull/4759/files#diff-ec6eb48af4ca0f0d2cac2569ba073a45015b0869fa1a5d9414f43d1472360812).

## Manual testing instructions

### Client-side rendering

Checkout this branch, build headless, and run:

`npm run dev -w @coveo/headless-commerce-react-samples`
 
Note that the `excludeDefaultParameters` option has been set to true in
the sample in both `src/pages/search-pager.tsx` and
`src/pages/product-listing-page.tsx`. To test that the default behavior
(i.e., including the default parameters) still works as before, just it
the option to `false` in both files.

### SSR

Checkout KIT-3462-d, build headless, and run:

`npm run dev -w @coveo/headless-ssr-commerce-samples`

Note that In SSR there's no way to include the default parameters (this
simply doesn't play well with SSR).