fix(atomic): fix accessibility issues with pager/results per page buttons (#4761)
This PR fixes the accessibility issues with the pager/results per page
buttons. More specifically, it was not possible to navigating between
pager or results per page buttons without selecting one, triggering a
refresh of the results.

The solution is to have the buttons inside a toolbar and implement the
logic described here:

https://www.w3.org/WAI/ARIA/apg/patterns/radio/#:~:text=For%20Radio%20Group%20Contained%20in%20a%20Toolbar

https://coveord.atlassian.net/browse/KIT-3767