fix(quantic): fixed issue with breadcrumb component when facet ids are used (#4791)
## [SFINT-5847](https://coveord.atlassian.net/browse/SFINT-5847)


### The issue

passing a facetId to a Quantic Facet component like the following:

```html
     <c-quantic-facet
        facet-id="obj"
        field="objecttype"
        label="Type"
        engine-id={engineId}
        data-cy="type"
      ></c-quantic-facet>
      <c-quantic-facet
        facet-id="file"
        display-values-as="link"
        field="filetype"
        label="File Type"
        engine-id={engineId}
      ></c-quantic-facet>
```

is causing an issue with the Quantic Breadcrumb component:
 
<img width="1455" alt="Screenshot 2024-11-28 at 11 27 45 AM"
src="https://github.com/user-attachments/assets/5a541704-c6e6-475a-9e8f-c32ebfd4746b"
/>


The issue is that the Quantic Breadcrumb component was reading the
facets from the store only using the facet `field` however the facet
`field` is used by the facets as `facetId` only when no facetId value is
passed to the facet components.


### After:
Updated the Quantic Breadcrumb component to read the facets info from
the store using the facet id instead of the facet field.

https://github.com/user-attachments/assets/56c3f3fa-b65b-4662-b3d1-4272e3879a8c



[SFINT-5847]:
https://coveord.atlassian.net/browse/SFINT-5847?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

---------

Co-authored-by: Etienne Rocheleau <erocheleau@coveo.com>