fix(headless-react): fix error when using keys other than their name for WithProps controllers (#4783)
https://coveord.atlassian.net/browse/KIT-3786

### Problem
Right now we have "cart" and "context" hard coded. This means that if
users use another key as "cart" for their cart implementation.

Lets say we have
`myCart: defineCart()
`

This condition would fail 
`      if ('cart' in controllers) {
`

### Solution
This PR fixes that problem by adding a _kind property to certain
controllers and then changing the condition to

`if (controller._kind == Kind.Cart) // where Kind is an enum`