fix(headless): handle end of stream message when answer can't be generated (#4800)
https://coveord.atlassian.net/browse/SVCC-4482

When perfoming a CRGA request, an "end of stream" Usage Analytics event
should be sent even when no answer is generated. However, when sending
the request to the Answer API, the "end of stream" Usage Analytics event
is sent only when an answer is generated.

This PR fixes the issue. In addition, tests have been added to cover the
handling of the various event source messages.