fix(headless): send end of stream event even when no answer is generated (#4801)
https://coveord.atlassian.net/browse/SVCC-4482

This is a backport of https://github.com/coveo/ui-kit/pull/4800 for v2.

When perfoming a CRGA request, an "end of stream" Usage Analytics event
should be sent even when no answer is generated. However, when sending
the request to the Answer API, the "end of stream" Usage Analytics event
is sent only when an answer is generated.

This PR fixes the issue. In addition, tests have been added to cover the
handling of the various event source messages.