chore: fix desynced package-lock (#4842)
KIT-282

Apparently there was a de sync between the package.json and
package-lock.json. We had different versions in the psons than in the
plock. I did a clean reset of the package-lock and reverted relay +
other dependencies back to their original versions.

For relay specifically, the new version of relay causes issues with
stencil made samples because of a node:crypto import. We will deal with
updating relay another time. This is blocking merges.