feat(SFINT-5848): Playwright migration for QuanticBreadcrumbManager tests (#4826)
SFINT-5848

We're now at 72 tests for Playwright, they run on my machine in 38.3s 🚀 

I also moved a bunch of the Cypress tests we had for the
QuanticBreadcrumbManager to unit tests. They mainly concerned the way we
"hide" breadcrumb values when they go over a threshold behind a "Show
more" button (no reason to test this via e2e as it's not even tied to
the size of the container, just a threshold value)

Same thing for the tests around different options such as a custom value
for the Show more threshold, and a custom delimiter for the Category
facets.

Other than that I added plenty of tests for each type of facets to make
sure the analytics Search event contains the correct format for the
value that was **de-selected** using the breadcrumb X.