chore: setup lit cohabitation (#4834)
First PR to scaffold the Lit migration, essentially #4787 without the
atomic-text migration.

Mainly:

## Web Dev Server
Replace the Stencil web dev server with a framework-agnostic one.
For now, it comes with degraded performance (it will need to be
replaced/adapted when the transition is over)

This comes with some hack removal in the Cypress tests.

## Autoloader
Setup an auto-loader for future Lit Elements, heavily inspired by
Shoelace

## Stencil Proxy
Proxy some files generated by Stencil to 'plug-in' what's relative to
Lit.

# Bikeshedding

- CSS: ought to setup CSS building for Lit
- Decorators & utils setup still