chore: relocate dist/components to dist/atomic/components (#4859)
With the Lit adaptation layer, we need the components directory in the
CDN files.
However, we only deploy dist/atomic to the CDN with the current
structure. Changing what we deploy to the CDN is the riskiest change
because even if we support only a given # of entry points, there is no
protection.
NPM and exports have this level of protection, so instead, I used the
"CDN" file structure as the immutable and changed things for NPM.

https://coveord.atlassian.net/browse/KIT-3858