fix(headless): don't leak pagination state to productSuggest call (#4667)
https://coveord.atlassian.net/browse/KIT-3660

For some context, 

Initially the ticket was "stop culling productSuggest to 5". It could
not be done since the pagination part of the state was overwriting the
perPage value for the productSuggest call (a feature from CAPI). So
let's say there was 25 products in the page, there would be 25 product
suggesting since both call were using the same pagination.

Pagination is useless for now in productSuggest so we can simply remove
it from the request.

---------

Co-authored-by: GitHub Actions Bot <>