docs(headless): updates the Headless typedoc uploading logic (#4876)
JIRA: https://coveord.atlassian.net/browse/KIT-3424

This PR changes the logic of how the Typedoc site is uploaded as an
artifact. This is to streamline the process of publishing the generated
docs at docs.coveo.com (the project that I'm working on these days).

I tested this approach in a local fork of this repo, you can look at the
latest run that uses the same logic as suggested in this PR:
https://github.com/npushkarskii/ui-kit/actions/runs/12832730326

----
> [!NOTE]
> * archiving in the workflow wasn't required as it resulted in creating
an archive inside an archive. GH automatically creates an archive when
we upload anything as an artifact.
>
> * the archive's names stays `headless-docs-{VERSION}.zip` (didn't
change)

The structure of the resulting archive will now be in the format: 

  1. The folder in the archive bears the name of the version.

  1. Inside the version folder, there are the generated files


![image](https://github.com/user-attachments/assets/2d53f93b-a564-4bb6-8984-768a4595f3ba)

---------

Co-authored-by: Alex Prudhomme <78121423+alexprudhomme@users.noreply.github.com>