feat(quantic): QuanticResultsPerPage fixes and refactor tests (#4889)
SFINT-5851

## IN THIS PR:
- We are migrating E2E tests for the `quanticResultsPerPage` component
from Cypress to Playwright and adding unit tests.
- We are also updating the behavior whenever an erroneous
`initialChoice` or no `initialChoice` is passed to the component. We
said
> "The initial selection for the number of result per page. This should
be part of the `choicesDisplayed` option. By default, this is set to the
first value in `choicesDisplayed`."

But in reality, we were always defaulting to the value `10` whenever no
value was passed,
And if the value was not within the `choicesDisplayed` we would
completely crash Quantic initialization by throwing an error.

Instead now, if the value is invalid (such as not a number, missing or
not part of the `choicesDisplayed`) we actually default to the first
value in `choicesDisplayed`. Log an error in the console, but do not
throw.

If the values in `choicesDisplayed` is invalid, we log an error in the
console, and display the componentError, but no longer throw an error.

## UNIT TESTS:

![image](https://github.com/user-attachments/assets/841ed67e-0a86-4e5b-92ca-27ac4968d1c4)


## E2E TESTS:

![image](https://github.com/user-attachments/assets/10cae2cb-398e-4859-a394-71248a1120e7)