chore: script to create new setup for atomic component (#4913)
https://coveord.atlassian.net/browse/KIT-3908

needs #4914 to work correctly so if you want to test it, test it from
there.

A new script to generate file structure and basic files for new lit
atomic components. It is used like this :


`npx nx run atomic:generate-component --name=atomic-ball
--output=src/components/search/atomic-ball`

The output arg is always relative to the atomic package root.


This is to help us during the migration and it will keep changing to
make it better.