fix(atomic): hide unhydrated results (#4900)
https://coveord.atlassian.net/browse/KIT-3775
Inspired by https://bbellmyers.github.io/2020/09/27/fouc-stencil.html.

Not a definitive fix: When Lit migration is done for the atomic-result,
we should validate for regression on that front, and also try to improve
the behavior even so slightly.
It might also be preferable to migrate all result components first, so
that we can "cut" this behavior neatly. (would likely not happen because
I _think_ we scan documentFragment from template element, so we should
have all the element defined by then (tho maybe not because of race
conditions, in which case
https://www.abeautifulsite.net/posts/flash-of-undefined-custom-elements
might be adapted for our needs)