chore(atomic): add index & lazy-index file in every use case (#4927)
https://coveord.atlassian.net/browse/KIT-3923

This PR adds the index.ts & lazy-index.ts file in every sub package of
atomic.

It was also required to change how some headless import worked in
insight & recommendations to clean up the index files.

@coveo/service-integration @coveo/service-core We need an index.ts & a
lazy-index.ts file for every sub package of atomic (commerce, insight,
ipx, ...) for the Lit migration and for going forward with Lit. This
will enable us to better organize the code and also make it possible for
implementers to only import a sub package specifically by doing `import
'@coveo/atomic/insight'` for example.

For insight and recommendations, there were special exports in the
index.ts file. I took the time to remove them and rework the imports in
the files affected. No functionality is lost, only small refactor.