fix(atomic): don't redirect when hovering a instant result & pressing Enter (#4938)
https://coveord.atlassian.net/browse/KIT-3910

We now have a new variable in suggestions-manager that tracks they
keyboard use. When someone press Enter when having a
suggestion/result/recent query selected with the keyboard, it will
trigger it's onClick event.
Whenever we stop using the keyboard and using the mouse it resets again
which will make the Enter event trigger the query in the search box
instead.
 
 I added a lot of tests to write in vitest.