test(quantic): fixed quantic timeframe facets e2e cypress test (#4972)
## [SFINT-5990](https://coveord.atlassian.net/browse/SFINT-5990)

### The issue: 

- In Quantic Timeframe Facet where there is a date input that leverages
the Salesforce date input component, when providing an invalid date,
salesforce displays a validation error message, that's expected. After
replacing the invalid date with a valid date, Salesforce used to
completely remove the validation error message element from the dom,
however now this has changed, I noticed that now they are just emptying
the content of this element and they mark it as `display: none` with
CSS.

### The solution: 
That was causing our test to fail as we try to count the number of
validation error message elements present in the page.
I modified our selectors to only select the VISIBLE validation error
message elements present in the page.

[SFINT-5990]:
https://coveord.atlassian.net/browse/SFINT-5990?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ