fix(headless): moves the docs notification step to the bottom (#4976)
In `release.yml`, there's a step responsible for notifying the docs repo
about the fact ot the release.

This PR moves the step to the bottom, after the Typedoc sites are
generated. Otherwise, the automation on docs's end fails because it
tries to fetch Typedoc artifacts that don't exist yet.