chore(atomic): update getAssetPath (#4946)
Update the logic to handle asset paths without the need of Stencil. The
logic is based on Stencil/core's `getAssetPath` function, where the
resourceUrl is dynamically determined based on how the app is built.

## Logic
When served from a CDN, the resourceUrl is computed based on the file
location (`import.meta.url`). This value is injected at build time via a
TypeScript custom transformer.
Otherwise, it is computed from the document location.
 
### Test Configuration Update:

*
[`packages/atomic/vitest.config.ts`](diffhunk://#diff-59124eddaa5017c075b045fd7028286720966faf8a9ff0f7a3807e08f23862c7R4-R13):
Updated the test configuration to define `import.meta.env.RESOURCE_URL`.

 
https://coveord.atlassian.net/browse/KIT-3944

---------

Co-authored-by: Louis Bompart <lbompart@coveo.com>