chore(atomic): remove useless FocusTargetController from pager components (#5002)
https://coveord.atlassian.net/browse/KIT-3993

This code simply did nothing. The only function called from the class
was setTarget and setTarget does nothing if no focus function is used
with it.

Initially I thought this was a bug but it is intended for this reason: 

We should refocus elswhere for facets but not for pagers because these
are not the same "use cases":

When you select facets, your intent is to filter the results and refine
the search.
When you select a page, your intent is to continue navigation.