feat(quantic): Allow RGA component to provide a custom "no answer" message when no answer is shown (#4919)
[SFINT-5926](https://coveord.atlassian.net/browse/SFINT-5926)

## IN THIS PR:

- Added the possibility for Quantic users to pass a slot with
`name=no-answer-message` to the `quanticGeneratedAnswer` component to
display a custom error message whenever CRGA could not return an answer.
If no slot is passed, we simply display nothing
- Covered this new logic with unit tests

### Notes about the solution:

- To display the custom error message, 2 conditions are required: 1) the
state `cannotAnswer` was returned as `true` and 2) the Quantic user has
passed a named slot in the component.
- We are using a hidden slot to detect whether a slot was passed to the
component, which allows us to check slot presence for the
`shouldDisplayNoAnswerMessage` check and display the no answer HTML
properly. This was the best way possible as LWC doesnt allow for
conditional rendering and as it keeps the additional logic to make this
feature work to a minimum in the JS file.


## DEMO:


https://github.com/user-attachments/assets/2cddabe7-5938-4fab-b23f-95012752da72



## TESTS: 
<img width="1040" alt="image"
src="https://github.com/user-attachments/assets/05554014-2bd8-4e03-bc84-04e9de998256"
/>

[SFINT-5926]:
https://coveord.atlassian.net/browse/SFINT-5926?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ