feat(headless commerce): re-implement field suggestions as faceted search (#5005)
https://coveord.atlassian.net/browse/KIT-3899

In the context of the Commerce API, the field suggestions functionality
is currently behind a feature flag.

In headless commerce, the feature was implemented essentially as "facet
value" search. E.g., if you enter "pa" in the search box, the field
suggestions would request facet values that match the wildcard facet
search query `*pa*` (e.g., "Clothing;**Pa**nts",
"Accessories;**Pa**ddles"). Selecting a value would then simply select
the corresponding facet value if available on the current page (if
selected from a standalone search box, you'd likely redirect to the full
search page or possibly to a listing page that matches the selected
category).

Here's an example of the "facet value" search behavior:


https://github.com/user-attachments/assets/61003f36-910c-4cbf-8ac7-2716ed81cc62

In reality, field suggestions should typically rather be used to provide
a "facet**ed**" search functionality. E.g., if you enter "jeans" in the
search box, field suggestions would return facet values that would
return results if selected when the search query is "pa" (e.g.,
"Clothing", "Clothing;Pants", "Clothing;Shorts"). Selecting a value
would set the search query to "jeans" and select the chosen facet value
(on the search page), or redirect to the search page with the correct
search params in a standalone search box scenario.

Here's an example of the "faceted" search behavior:


https://github.com/user-attachments/assets/77997be7-4de5-4e3e-8afa-62bfc36e01fc

This pull requests re-implements headless commerce field suggestions in
the latter fashion.

An important caveat: we do have one live customer using the feature as
it was originally designed, so we need to introduce this change in a
non-breaking manner for them. The customer is leveraging the feature as
a hack to implement a "listing page search" functionality (i.e., provide
`ec_category` field suggestions and redirect to the matching listing
page when a suggestion is selected). Therefore, we must not break the
`CategoryFieldSuggestions` controller. We can remove support for the
`FieldSuggestions` controller altogether, though, as it's not used by
any live customers.

For this reason, the re-implemented headless controllers are introduced
with a new name ("filter suggestions").