chore(atomic): add injectStylesForNoShadowDOM decorator (#5037)
### Description
Added the `@injectStylesForNoShadowDOM(styles)` decorator for components
that render directly into the DOM. This decorator is useful for
components with styles that need to be added but are not encapsulated in
a shadow DOM.

When using this decorator, the component's styles will be injected into
the closest shadow root (if the component is placed in a parent
element's shadow DOM). If not, the styles will be placed in the document
header.

If multiple similar components are placed in the same parent shadow root
(e.g., multiple check icons in the facet element), only one style tag
will be added to the facet's shadow root.

### Usage
```ts
const styles = css`
  div {
    display: block;
    color: red;
  }
`;

@injectStylesForNoShadowDOM(styles)
@customElement('styled-element')
class StyledElement extends LitElement {
  render() {
    return html`<div>Hello World</div>`;
  }
}
```

https://coveord.atlassian.net/browse/KIT-4032