chore: split release:phase1 & release:phase3 in two phases (#5042)
https://coveord.atlassian.net/browse/KIT-4035

For the CDN publication we need to bump only and not publish to npm. We
also dont want the root package json to be bumped. The current setup
does not allow us to do this.
this is the new phases : 
```
    # phase0 is locking master
    # phase1 is the versions bump
    # phase2 is the root package json bump
    # phase3 is the publish
    # phase4 is the reify
    # phase5 is the commit version bumps
```


So for prerelease npm we would do 
```
phase1
phase3
phase4
```

for prerelease cdn we would do
```
phase1
phase4
```

for full release we would do 
```
phase0
phase1
phase2
phase3
phase4
phase5
```

Not sure if it will work 😄 . Lets try it !