ci: fix failing quantic:release:phase1 during prerelease (#5064)
https://coveord.atlassian.net/browse/KIT-4049

Disabling DEBUG in that command will fix the failing prerelease since
quantic won't have to print over `80mb` of text during it's build:doc
phase 😆 . I am not sure why but it ends up exiting before the text can
print. Not sure why but we definitely don't need to enable DEBUG for
this.

I tested it with a custom github action and confirmed it was passing
without `DEBUG: '*'` and was failing with it.