ci: deploy the CDN on dev on each commit (#5045)
https://coveord.atlassian.net/browse/KIT-3962

### You should review #5042 and #5041 before the one here and the one in
ui-kit-cd

with https://github.com/coveo-platform/ui-kit-cd/pull/66, it should
theoretically work.

On each commit we will create a new deployment package for dev-stg-prd
but we won't approve it to go to stg.

On tuesdays, we will begin a standard deployment for dev-stg-prd, it
will go to dev, we will approve it to go to stg and then leave it like
that until wednesday where we will approve it to go to prd.