chore: revert auto merge all dependencies (#5066)
https://coveord.atlassian.net/browse/KIT-282

The reason for this is that currently, [renovate groups a patch of
tailwindcss with the major-all
PR](https://github.com/coveo/ui-kit/pull/5057). I suspect that the
current groups are the root cause. I put it back to how it was.


Anyway, the automerge stuff was not working and it seems like there is
more setup needed to do on the Coveo side to enable this. We can do it
properly another time.