fix(quantic): fixed issue with custom facet captions inside the refine modal (#5075)
## [SFINT-6003](https://coveord.atlassian.net/browse/SFINT-6003)

## The issue: 

When facets are registered in the Quantic Store, the data about captions
is not being included. This leads to the following behavior:

In Insight, where facets are solely displayed via the Refine Modal,
captions do not appear.

In Search, captions work as expected when the screen is wide, but if the
screen is narrowed and facets move into the Refine Modal, captions are
lost.


The example is using a caption of `Text file` for the value `txt` and a
caption `DZA` for the value `Algeria`.





https://github.com/user-attachments/assets/9483ec2e-bb38-4c61-a368-cee6699b47b5






## The solution:

The custom facet captions are passed when a facet registers to the
Quantic Store, the Quantic Refine Modal content component will then take
care of displaying the captions in the refine modal as expected.


## Demo:


https://github.com/user-attachments/assets/13718a9e-3fc0-4a17-8ea1-f371d45d340b



[SFINT-6003]:
https://coveord.atlassian.net/browse/SFINT-6003?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ