fix(quantic): custom no-answer message displayed when search returns no results (#5084)
[SFINT-6041](https://coveord.atlassian.net/browse/SFINT-6041)

## ISSUE:

The custom no-answer message slot was being displayed when the search
returned no results. So essentially, it would display the generated
answer card with the quanticNoResult component.

<img width="706" alt="image"
src="https://github.com/user-attachments/assets/f0495877-ff83-41af-b511-f0bf17d14d55"
/>

The issue was because we didn't check for the presence of results in the
condition to display this custom message.

## SOLUTION:

We changed the condition to add a check for the presence of results
before rendering the custom no answer message.

Also added unit tests for that.

## DEMO:


https://github.com/user-attachments/assets/fbafdb0c-b374-4eb6-8307-f96f6d0dca5d

## TESTS:

<img width="1042" alt="image"
src="https://github.com/user-attachments/assets/1f557168-a359-406a-bedf-6213bc8c80e5"
/>





[SFINT-6041]:
https://coveord.atlassian.net/browse/SFINT-6041?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ