fix(quantic): handle corner-cases for ResultsPerPage (#5096)
SFINT-5954

Handle corner-cases for ResultsPerPage.

Added unit tests for those corner cases:

1. If you set the `initialChoice` with a negative value, or with not a
number. There will be a console error and the component won't initialize
but instead display the QuanticError.
2. If you provide a negative choicesToDisplay value, there will be a
console error and the component won't initialize but instead display the
QuanticError.


![image](https://github.com/user-attachments/assets/78e9298c-586a-4ad0-b3a9-9c92ab08e7ef)


3. If the `initialChoice` provided is not part of the
`choicesToDisplay`, a console warning (instead of a console error
previously) will be displayed and we will default to the first value in
the `choicesToDisplay`

---------

Co-authored-by: Louis Bompart <lbompart@coveo.com>