ci: more explicit nx build cache (#5051)
https://coveord.atlassian.net/browse/KIT-4031


Our nx setup was overly complicated. I simplified it a lot. Below I
explain what `inputs` and `outputs` mean for NX and also explain the
changes.

## What changes
In nx the `inputs` array are files that would trigger a new build if
they change. The default is every non-`outputs` file.
The `outputs` array are files that are generated by the build and are
then cached.

- Before, we had a `cached:build` command that was taking care of
caching `build`, from now on, we will simply just cache `build`. Having
a separate command for that is unnecessary.
- Before we were defining some special `inputs`. From now on, there
won't be any special `inputs` and now every project will rebuild if any
file in its directory changes.
- Before, we were also defining `outputs` in the nx file that were being
reused in every project. From now on, we will explicitly define every
`outputs` that is being cached and will add "cache": true.

## Example with headless-ssr-commerce
### Before
Before this was the project.json, "build" depended on "cached:build",
`cached:build` was `cache:true`, it had special inputs and outputs.
`build` had the
[production](https://github.com/coveo/ui-kit/blob/0bb6ae688ed45139dc67cdb21e2691d6ea741e96/nx.json#L48)
input. It was not explicit that it was caching the .next folder since
this was hidden in the `nx.json` file.
`project.json`
```
"cached:build": {
      "executor": "nx:run-commands",
      "options": {
        "command": "npm run-script -w=@coveo/release if-not-cdn -- \"npm run build:next\"",
        "cwd": "packages/samples/headless-ssr-commerce"
      }
    },
    "build": {
      "dependsOn": ["cached:build"],
      "executor": "nx:noop"
    }
```
`nx.json`
```
"build": {
      "inputs": ["^production", "production"],
      "dependsOn": ["^build"]
    },
    "cached:build": {
      "inputs": ["^production", "production"],
      "outputs": [
        "{projectRoot}/dist",
        "{projectRoot}/www",
        "{projectRoot}/temp",
        "{projectRoot}/build",
        "{projectRoot}/.angular",
        "{projectRoot}/.next",
        "{projectRoot}/next-env.d.ts"
      ],
      "dependsOn": ["^build"],
      "cache": true
    },
```

### After
After, the `project.json` explicitly explains that `.next` is cached.
And the `nx.json`'s `build` command simply calls every other `build`
command in the repo.
`project.json`
```
 "build": {
      "executor": "nx:run-commands",
      "outputs": ["{projectRoot}/.next", "{projectRoot}/next-env.d.ts"],
      "cache": true,
      "options": {
        "command": "npm run-script -w=@coveo/release if-not-cdn -- \"npm run build:next\"",
        "cwd": "{projectRoot}"
      }
    }
```
`nx.json`
```
    "build": {
      "dependsOn": ["^build"]
    },
```