fix(quantic): added name attribute to the Quantic tab component (#5130)
## [SFINT-5984](https://coveord.atlassian.net/browse/SFINT-5984)

## The issue
Previously, the label property of the QuanticTab component was used as
the unique identifier for the tab in the Headless state. As a result,
the label value:

- Was used as the tab parameter in the URL
- Was sent as the tab value in Search API requests
- Was included in the analytics payload for events related to tab
interactions

However, this behavior is not ideal. The label is primarily meant for
display purposes and should support localization. For example, "Issues"
might be translated to "problemes" which would then appear in the URL,
analytics, and backend pipelines — potentially leading to
inconsistencies or issues in ML models, reporting, or pipeline rules.

<img width="1724" alt="Screenshot 2025-04-07 at 8 51 43 AM"
src="https://github.com/user-attachments/assets/c07e2c2a-ece4-4149-be4c-557b33f9eba1"
/>


## The solution
To address this, I added a new public property to the QuanticTab
component: `name`. I didn't go with `tabId` because `name` was use by
the atomic equivalent component.

The new name property is now used as the unique tab identifier.

- It appears in the URL
- It is sent to the Search API
- It is logged in analytics payloads

The label remains for UI display only and can be freely localized
without affecting functionality.

This update ensures a clean separation between the user-facing display
and the technical identifiers used internally and externally.

<img width="1728" alt="Screenshot 2025-04-07 at 8 50 07 AM"
src="https://github.com/user-attachments/assets/0c0fb384-24db-4819-bea3-a0aab285e26f"
/>


[SFINT-5984]:
https://coveord.atlassian.net/browse/SFINT-5984?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ