fix(quantic): breadcrumb value not displayed for date and numeric filters (#5145)
## [SFINT-6087](https://coveord.atlassian.net/browse/SFINT-6087)

### The issue:
Breadcrumb values were not displayed correctly when a numeric filter or
a date filter were selected.
<img width="662" alt="Screenshot 2025-04-07 at 4 08 52 PM"
src="https://github.com/user-attachments/assets/ffbf6195-8835-452d-abfb-430bfc0c5276"
/>

this is because the filters are registered in the Headless state always
with the facet id of the facet that owns the filter in addition to the
suffix `_input`:

https://github.com/coveo/ui-kit/blob/bbb30961fedc4168dcae81aad561f5c864d7c4a5/packages/quantic/force-app/main/default/lwc/quanticNumericFacet/quanticNumericFacet.js#L340-L344

https://github.com/coveo/ui-kit/blob/bbb30961fedc4168dcae81aad561f5c864d7c4a5/packages/quantic/force-app/main/default/lwc/quanticTimeframeFacet/quanticTimeframeFacet.js#L490-L492

However, in the Quantic store, only the facet id of the facet that owns
the filter is stored.


### The solution:

In order to get the necessary data to display the breadcrumbs for a
filter, we need to access the data of the facet that owns the filter an
not the filter directly, that's why I added a logic always removes the
suffix `_input` from the key we use to access the store, this way when
we need the data of the filter `exampleField_inout`, we would get the
data from the facet that owns this filter, aka the facet with the id
`exampleField` .


https://github.com/user-attachments/assets/88bb41a6-0923-44ba-8c0a-dbe565afa57d




[SFINT-6087]:
https://coveord.atlassian.net/browse/SFINT-6087?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ